+++ Adj/Program/Option.hs
@@ -1,13 +1,13 @@
 module Adj.Program.Option where
 
-import Adj.Algebra.Category ((:+:) (Option, Adoption), Terminal (Terminal))
+import Adj.Algebra.Category ((:+:) (This, That), Terminal (Terminal))
 
 type Option = (:+:) Terminal
 
 pattern Some :: a -> Option a
-pattern Some x <- Adoption x
-	where Some x = Adoption x
+pattern Some x <- That x
+	where Some x = That x
 
 pattern None :: Option a
-pattern None <- Option Terminal
-	where None = Option Terminal
+pattern None <- This Terminal
+	where None = This Terminal